Card Element that displays in the card swiper
Pretty similar to the older version of this element, except it has less funcionality and thus more simple.
